refactor stats
