+++ src/main.js
@@ -1,10 +1,9 @@
 import Search from "./components/search";
 import Profile from "./components/profile";
 import Menu from "./components/menu";
-import Statistic from "./components/statistic";
 import Footer from "./components/footer";
 import filmCards from "./data/cards";
-import {render} from './components/utils';
+import {render} from './utils';
 import Page from "./controllers/page";
 import SearchController from "./controllers/serarch";
 import StatsController from "./controllers/stats";
