+++ src/utils.js
@@ -1,3 +1,5 @@
+import data from './data/cards.js';
+
 const KEY_CODE_ESCAPE = 27;
 const KEY_CODE_ENTER = 13;
 const NUMBER_SHOW_FILMS = 5;
@@ -36,6 +38,23 @@ const unrender = (element) => {
   }
 };
 
+const getNumberFilmsStat = (control) => data.reduce((count, film) => (film.controls[control] ? count + 1 : count), 0);
+
+const stats = {
+  get allFilmsNumber() {
+    return data.length;
+  },
+  get watchlistNumber() {
+    return getNumberFilmsStat(`isAddedToWatchlist`);
+  },
+  get historyNumber() {
+    return getNumberFilmsStat(`isMarkedAsWatched`);
+  },
+  get favoritesNumber() {
+    return getNumberFilmsStat(`isFavorite`);
+  }
+};
+
 export {
   createElement,
   render,
@@ -44,5 +63,6 @@ export {
   KEY_CODE_ENTER,
   NUMBER_SHOW_FILMS,
   NUMBER_SHOW_TOP_RATED_FILMS,
-  NUMBER_SHOW_MOST_COMMENTED_FILMS
+  NUMBER_SHOW_MOST_COMMENTED_FILMS,
+  stats
 };
